# Pure `HashAndNibblize` function