--- a/erigon-lib/commitment/hex_patricia_hashed_bench_test.go
+++ b/erigon-lib/commitment/hex_patricia_hashed_bench_test.go
@@ -49,7 +49,7 @@ func Benchmark_HexPatriciaHashed_Process(b *testing.B) {
 	require.NoError(b, err)
 
 	hph := NewHexPatriciaHashed(length.Addr, ms, ms.TempDir())
-	upds := WrapKeyUpdates(b, ModeDirect, HashAndNibblizeKey, nil, nil)
+	upds := WrapKeyUpdates(b, ModeDirect, KeyToHexNibbleHash, nil, nil)
 	defer upds.Close()
 
 	b.ResetTimer()