--- a/erigon-lib/commitment/utils.go
+++ b/erigon-lib/commitment/utils.go
@@ -1,6 +1,9 @@
 package commitment
 
 import (
+	"errors"
+	"fmt"
+
 	"github.com/erigontech/erigon-lib/common/length"
 	ecrypto "github.com/erigontech/erigon-lib/crypto"
 )
@@ -16,5 +19,120 @@ func HashAndNibblizeKey(key []byte) []byte {
 		storageKeyHash := ecrypto.Keccak256(key[length.Addr:])
 		compactedHashKey = append(compactedHashKey, storageKeyHash...)
 	}
-	return nibblize(compactedHashKey)
+	return Nibblize(compactedHashKey)
+}
+
+func hexToCompact(key []byte) []byte {
+	zeroByte, keyPos, keyLen := makeCompactZeroByte(key)
+	bufLen := keyLen/2 + 1 // always > 0
+	buf := make([]byte, bufLen)
+	buf[0] = zeroByte
+	return decodeKey(key[keyPos:], buf)
+}
+
+func makeCompactZeroByte(key []byte) (compactZeroByte byte, keyPos, keyLen int) {
+	keyLen = len(key)
+	if hasTerm(key) {
+		keyLen--
+		compactZeroByte = 0x20
+	}
+	var firstNibble byte
+	if len(key) > 0 {
+		firstNibble = key[0]
+	}
+	if keyLen&1 == 1 {
+		compactZeroByte |= 0x10 | firstNibble // Odd: (1<<4) + first nibble
+		keyPos++
+	}
+
+	return
+}
+
+func decodeKey(key, buf []byte) []byte {
+	keyLen := len(key)
+	if hasTerm(key) {
+		keyLen--
+	}
+	for keyIndex, bufIndex := 0, 1; keyIndex < keyLen; keyIndex, bufIndex = keyIndex+2, bufIndex+1 {
+		if keyIndex == keyLen-1 {
+			buf[bufIndex] = buf[bufIndex] & 0x0f
+		} else {
+			buf[bufIndex] = key[keyIndex+1]
+		}
+		buf[bufIndex] |= key[keyIndex] << 4
+	}
+	return buf
+}
+
+func CompactedKeyToHex(compact []byte) []byte {
+	if len(compact) == 0 {
+		return compact
+	}
+	base := keybytesToHexNibbles(compact)
+	// delete terminator flag
+	if base[0] < 2 {
+		base = base[:len(base)-1]
+	}
+	// apply odd flag
+	chop := 2 - base[0]&1
+	return base[chop:]
+}
+
+func keybytesToHexNibbles(str []byte) []byte {
+	l := len(str)*2 + 1
+	var nibbles = make([]byte, l)
+	for i, b := range str {
+		nibbles[i*2] = b / 16
+		nibbles[i*2+1] = b % 16
+	}
+	nibbles[l-1] = 16
+	return nibbles
+}
+
+// hasTerm returns whether a hex key has the terminator flag.
+func hasTerm(s []byte) bool {
+	return len(s) > 0 && s[len(s)-1] == 16
+}
+
+func commonPrefixLen(b1, b2 []byte) int {
+	var i int
+	for i = 0; i < len(b1) && i < len(b2); i++ {
+		if b1[i] != b2[i] {
+			break
+		}
+	}
+	return i
+}
+
+func Nibblize(key []byte) []byte { // nolint:unused
+	nibblized := make([]byte, len(key)*2)
+	for i, b := range key {
+		nibblized[i*2] = (b >> 4) & 0xf
+		nibblized[i*2+1] = b & 0xf
+	}
+	return nibblized
+}
+
+// compactKey takes a slice of nibbles and compacts them into the original byte slice.
+// It returns an error if the input contains invalid nibbles (values > 0xF).
+func compactKey(nibbles []byte) ([]byte, error) {
+	// If the number of nibbles is odd, you might decide to handle it differently.
+	// For this example, we'll return an error.
+	if len(nibbles)%2 != 0 {
+		return nil, errors.New("nibbles slice has an odd length")
+	}
+
+	key := make([]byte, len(nibbles)/2)
+	for i := 0; i < len(key); i++ {
+		highNibble := nibbles[i*2]
+		lowNibble := nibbles[i*2+1]
+
+		// Validate that each nibble is indeed a nibble
+		if highNibble > 0xF || lowNibble > 0xF {
+			return nil, fmt.Errorf("invalid nibble at position %d or %d: 0x%X, 0x%X", i*2, i*2+1, highNibble, lowNibble)
+		}
+
+		key[i] = (highNibble << 4) | (lowNibble & 0x0F)
+	}
+	return key, nil
 }