This PR extracts `HashAndNibblize` as a utility function, not dependent on `HexPatriciaTrieHashed` . 

Doing some benchmark comparisons this change doesn't impact performance (in fact it looks to make an improvement).

These are the comparisons:

- For 25 million keys:
  - (before) `hph.HashAndNibblizeKey`: 136.664s
  - (after) pure `HashAndNibblizeKey` : 108.215s

- For 35 million keys:
  - (before) `hph.HashAndNibblizeKey`: 351.598s
  - (after) pure `HashAndNibblizeKey` :   326.218s

- For 50 million keys:
  - (before) `hph.HashAndNibblizeKey`: 1055.217s
  - (after) pure `HashAndNibblizeKey` :   568.075s